requirements.txt
fastapi['standard']==0.128.0
sqlmodel==0.0.31
alembic==1.18.1
bcrypt==5.0.0
python-jose==3.5.0